Fixed: Average timings would previously sometimes be incorrect since they took into account unset values of the timing history.
Other changes:
* Timer uses Java's TimeUnit for conversions now.
* Timer and Timings use a long counter now.
* Minor internal Javadoc clarifications.

+++ src/main/java/com/nisovin/shopkeepers/util/timer/Timings.java
@@ -4,7 +4,7 @@ public interface Timings {
 
 	public void reset();
 
-	public int getCounter();
+	public long getCounter();
 
 	public double getAverageTimeMillis();
